+++ src/providers/Completion/GMmetaCompletionProvider.ts
@@ -14,7 +14,7 @@ const GMmetaCompletionProvider = vscode.languages.registerCompletionItemProvider
                         doc.appendMarkdown(i.value);
                         break;
                     case 'code':
-                        doc.appendCodeblock(i.value, 'javascript');
+                        doc.appendCodeblock(i.value, 'typescript');
                         break;
                     default:
                         doc.appendText(i.value);