+++ CHANGELOG.md
@@ -2,7 +2,17 @@
 
 ## [Unreleased]
 
-- Nothing
+## [0.0.2] - 2022-07-29
+
+### Added
+
+- Support highlight `GM.notification`.
+- Support metadata completion `@exclude`, `@exclude-match`.
+- Support metadata `@installURL`, `@contributionURL`, `@contributionAmount`, `@compatible`, `@incompatible`.
+
+### Changed
+
+- `*` in `@connect` tag will be highlighted.
 
 ## [0.0.1] - 2022-07-25